+++ app/internal/modules/file-station/factory/options.go
@@ -3,7 +3,7 @@ package factory
 import (
 	"github.com/mondegor/go-storage/mrstorage"
 	"github.com/mondegor/go-webcore/mrcore"
-	"github.com/mondegor/go-webcore/mrserver"
+	"github.com/mondegor/go-webcore/mrserver/mrparser"
 	"github.com/mondegor/go-webcore/mrserver/mrresponse"
 	"github.com/mondegor/go-webcore/mrtool"
 )
@@ -12,7 +12,7 @@ type (
 	Options struct {
 		Logger         mrcore.Logger
 		ServiceHelper  *mrtool.ServiceHelper
-		RequestParser  mrserver.RequestParserPath
+		RequestParser  *mrparser.String
 		ResponseSender *mrresponse.Sender
 
 		UnitImageProxy *UnitImageProxyOptions