+++ app/internal/modules/provider-accounts/usecase/provider-account-api/company_page_logo_interface.go
@@ -9,7 +9,7 @@ import (
 
 type (
 	CompanyPageLogoService interface {
-		StoreFile(ctx context.Context, accountID mrtype.KeyString, image mrtype.File) error
+		StoreFile(ctx context.Context, accountID mrtype.KeyString, image mrtype.Image) error
 		RemoveFile(ctx context.Context, accountID mrtype.KeyString) error
 	}